# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RingiFlow API
  description: ワークフロー管理システム RingiFlow の BFF API
  license:
    name: MIT
    identifier: MIT
  version: 0.1.0
paths:
...
  /api/v1/auth/csrf:
    get:
      tags:
      - auth
      summary: GET /api/v1/auth/csrf
      description: |-
        CSRF トークンを取得する。
        セッションが存在しない場合は新規作成し、存在する場合は既存のトークンを返す。
      operationId: csrf
      responses:
        '200':
          description: CSRF トークン
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfResponseData'
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: POST /api/v1/auth/login
      description: |-
        メール/パスワードでログインし、セッションを確立する。

        ## 認証フロー

        1. Core API でユーザーを検索（`GET /internal/users/by-email`）
        2. Auth Service でパスワードを検証（`POST /internal/auth/verify`）
        3. セッションを作成し Cookie を設定

        ## タイミング攻撃対策

        ユーザーが存在しない場合も Auth Service にダミーリクエストを送信し、
        処理時間を均一化してユーザー存在確認攻撃を防ぐ。

        ## ヘッダー

        - `X-Tenant-ID`: テナント ID（必須）

        ## リクエストボディ

        ```json
        {
          "email": "user@example.com",
          "password": "password123"
        }
        ```
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseData'
        '401':
          description: 認証失敗
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '503':
          description: サービス利用不可
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/auth/logout:
    post:
      tags:
      - auth
      summary: POST /api/v1/auth/logout
      description: セッションを無効化してログアウトする。
      operationId: logout
      responses:
        '204':
          description: ログアウト成功
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: GET /api/v1/auth/me
      description: 現在のユーザー情報と権限を取得する。
      operationId: me
      responses:
        '200':
          description: ユーザー情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponseData'
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
...
  /api/v1/users:
    get:
      tags:
      - users
      summary: GET /api/v1/users
      description: |-
        テナント内のユーザー一覧を取得する。
        ステータスフィルタに対応（省略時は deleted 以外すべて）。
      operationId: list_users
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: ユーザー一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserItemData'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
    post:
      tags:
      - users
      summary: POST /api/v1/users
      description: |-
        ユーザーを作成する。

        ## フロー

        1. 入力バリデーション
        2. 初期パスワード生成
        3. Core Service でユーザー作成
        4. Auth Service で認証情報作成
        5. 初期パスワード付きレスポンス返却
      operationId: create_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        '201':
          description: ユーザー作成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponseData'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: メールアドレス重複
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
  /api/v1/users/{display_number}:
    get:
      tags:
      - users
      summary: GET /api/v1/users/{display_number}
      description: 表示用連番でユーザー詳細を取得する。
      operationId: get_user_detail
      parameters:
      - name: display_number
        in: path
        description: ユーザー表示番号
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: ユーザー詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailData'
        '404':
          description: ユーザーが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
    patch:
      tags:
      - users
      summary: PATCH /api/v1/users/{display_number}
      description: ユーザー情報を更新する（名前、ロール）。
      operationId: update_user
      parameters:
      - name: display_number
        in: path
        description: ユーザー表示番号
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
        required: true
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseData'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: ユーザーが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
  /api/v1/users/{display_number}/status:
    patch:
      tags:
      - users
      summary: PATCH /api/v1/users/{display_number}/status
      description: |-
        ユーザーステータスを変更する。
        セッションから自身の user_id を取得し、Core Service に渡す。
      operationId: update_user_status
      parameters:
      - name: display_number
        in: path
        description: ユーザー表示番号
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserStatusRequest'
        required: true
      responses:
        '200':
          description: ステータス変更成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseData'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: ユーザーが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
...
components:
  schemas:
...
    CreateUserRequest:
      type: object
      description: ユーザー作成リクエスト
      required:
      - email
      - name
      - role_id
      properties:
        email:
          type: string
        name:
          type: string
        role_id:
          type: string
          format: uuid
    CreateUserResponseData:
      type: object
      description: ユーザー作成レスポンス（初期パスワード付き）
      required:
      - id
      - display_id
      - display_number
      - name
      - email
      - role
      - initial_password
      properties:
        id:
          type: string
        display_id:
          type: string
        display_number:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        role:
          type: string
        initial_password:
          type: string
...
    CsrfResponseData:
      type: object
      description: CSRF トークンデータ
      required:
      - token
      properties:
        token:
          type: string
...
    LoginRequest:
      type: object
      description: ログインリクエスト
      required:
      - email
      - password
      properties:
        email:
          type: string
        password:
          type: string
    LoginResponseData:
      type: object
      description: ログインレスポンスデータ
      required:
      - user
      properties:
        user:
          $ref: '#/components/schemas/LoginUserResponse'
    LoginUserResponse:
      type: object
      description: ログインユーザー情報
      required:
      - id
      - email
      - name
      - tenant_id
      - roles
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        tenant_id:
          type: string
          format: uuid
        roles:
          type: array
          items:
            type: string
    MeResponseData:
      type: object
      description: 現在のユーザー情報データ
      required:
      - id
      - email
      - name
      - tenant_id
      - tenant_name
      - roles
      - permissions
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        tenant_id:
          type: string
          format: uuid
        tenant_name:
          type: string
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
...
    ProblemDetails:
      type: object
      description: |-
        エラーレスポンス（RFC 9457 Problem Details）

        すべてのサービスで統一されたエラーレスポンス形式。
        `type` フィールドは URI で問題の種類を識別する。
      required:
      - type
      - title
      - status
      - detail
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
          minimum: 0
        detail:
          type: string
...
    UpdateUserRequest:
      type: object
      description: ユーザー更新リクエスト
      properties:
        name:
          type:
          - string
          - 'null'
        role_id:
          type:
          - string
          - 'null'
          format: uuid
    UpdateUserStatusRequest:
      type: object
      description: ユーザーステータス変更リクエスト
      required:
      - status
      properties:
        status:
          type: string
...
    UserDetailData:
      type: object
      description: ユーザー詳細レスポンス
      required:
      - id
      - display_id
      - display_number
      - name
      - email
      - status
      - roles
      - permissions
      - tenant_name
      properties:
        id:
          type: string
        display_id:
          type: string
        display_number:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        status:
          type: string
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
        tenant_name:
          type: string
    UserItemData:
      type: object
      description: ユーザー一覧の要素データ
      required:
      - id
      - display_id
      - display_number
      - name
      - email
      - status
      - roles
      properties:
        id:
          type: string
        display_id:
          type: string
        display_number:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        status:
          type: string
        roles:
          type: array
          items:
            type: string
...
    UserResponseData:
      type: object
      description: ユーザー簡易レスポンス（更新・ステータス変更用）
      required:
      - id
      - name
      - email
      - status
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        status:
          type: string
...
  securitySchemes:
    session_auth:
      type: apiKey
      in: cookie
      name: session_id
...